Extract outcome review router models
